Move Command selector and home files to correct directory;
