+++ app/layout.tsx
@@ -1,10 +1,11 @@
-import { CommandMenu } from '@/components/command';
-import { CommandMenuProvider } from '@/components/command/command-menu-context';
 import InteractiveConfigContextWrapper from '@/components/interactive/ContextWrapper';
+import { CommandMenu } from '@/components/interactive/Selectors/Command';
+import { CommandMenuProvider } from '@/components/interactive/Selectors/Command/command-menu-context';
 import Head from '@/components/jrg/appwrapper/Head';
 import { SidebarContentProvider } from '@/components/jrg/appwrapper/SidebarContentManager';
 import { SidebarContext } from '@/components/jrg/appwrapper/SidebarContext';
 import { SidebarMain } from '@/components/jrg/appwrapper/SidebarMain';
+import { SolanaWalletProvider } from '@/components/jrg/wallet/wallet-provider';
 import '@/components/jrg/zod2gql/zod2gql';
 import { SidebarProvider } from '@/components/ui/sidebar';
 import { Toaster } from '@/components/ui/toaster';
@@ -14,7 +15,6 @@ import { cookies } from 'next/headers';
 import { ReactNode } from 'react';
 import './globals.css';
 import { metadata, viewport } from './metadata';
-import { SolanaWalletProvider } from '@/components/jrg/wallet/wallet-provider';
 
 const inter = Inter({ subsets: ['latin'] });
