+++ components/interactive/Home/call-to-action.tsx
@@ -1,7 +1,6 @@
 'use client';
-import React from 'react';
 import Link from 'next/link';
-import { Button } from '../ui/button';
+import { Button } from '../../ui/button';
 
 export function CallToAction() {
   return (
